DB 마이그레이션 도구 Alembic 도입 : refactor : 기존 테이블 id BigInteger 통일 (players, hall_of_fame) https://github.com/PickerPicker/PickerPicker/issues/134

+++ backend/src/models/hall_of_fame.py
@@ -2,7 +2,7 @@
 역대 1위 기록. ended_at IS NULL = 현재 챔피언 (항상 최대 1개).
 """
 from datetime import datetime
-from sqlalchemy import String, Integer, DateTime
+from sqlalchemy import String, Integer, BigInteger, DateTime
 from sqlalchemy.orm import Mapped, mapped_column
 from src.core.database import Base
 
@@ -10,7 +10,7 @@ from src.core.database import Base
 class HallOfFame(Base):
     __tablename__ = "hall_of_fame"
 
-    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
     nickname: Mapped[str] = mapped_column(String(50), nullable=False)
     score: Mapped[int] = mapped_column(Integer, nullable=False)
     started_at: Mapped[datetime] = mapped_column(DateTime, nullable=False)
